protocol/flags: fix clippy warning from 1.42 toolchain build
This commit fixes 'trivially-copy-pass-by-ref' lint, which was enabled
in 1.42 by default. It was downgraded to `pedantic` level and disabled
by default in later versions of clippy.

I am still accepting this recommendation cause this was the first
edition of the function, and the reason for adding the reference was for
symmetry with the set_* functions only.