error: coding: normalize the RsDnsError enum
Initially I wanted to split protocol errors to a dedicated enum.
However, that would require definition of a separate Result.

I decided to leave the single errors enum for convenience. Most of the
errors in this enum will be protocol related. There will be small amount
of wrapper errors for the net module.