ci: bump MSRV to 1.46
async-std started failing to compile on 1.45 due to a new indirect
dependency 'socket2' which in version 0.4.0 fails with the following
error:

   Compiling socket2 v0.4.0
error[E0658]: `match` is not allowed in a `const fn`
   --> /home/runner/.cargo/registry/src/github.com-1ecc6299db9ec823/socket2-0.4.0/src/lib.rs:156:9
    |
156 | /         match address {
157 | |             SocketAddr::V4(_) => Domain::IPV4,
158 | |             SocketAddr::V6(_) => Domain::IPV6,
159 | |         }
    | |_________^
    |
    = note: see issue #49146 <https://github.com/rust-lang/rust/issues/49146> for more information

   Compiling signal-hook-registry v1.3.0
error: aborting due to previous error

For more information about this error, try `rustc --explain E0658`.
error: could not compile `socket2`.