Revert "proto: move enums to rust naming convention"
I am reverting this commit because Rust naming convention for enum items
simply make these constants something they are not. These names are so
common in DNS lingo, that it is impossible to name them differently.

Bump of MSRV to 1.51 is required as this is the first version that
supports disabling clippy::upper_case_acronyms lint.

This reverts commit f4869d059d6e8967bfb99666ab91e1aa23e08d34.