net: fix clippy::manual_flatten error
In the beta build clippy complains on manual_flatten. Iterator::flatten
crates an iterator which removes a single level of indirection. I.e. if
iterated values are iterators or can be converted to iterators, those
are flattened into a single iterator. As Result and Option values can be
converted to an iterator, it makes sence to replace the verbose "if let
Ok(_) =" construct into a call to flatten.