errors: remove CnameChainTooLong error
This error is not needed, as the limitation chain length. This is
because the chain can never be longer than the number of records in a
response. There is no risk of an infinite loop, as every time a CNAME is
found, it is extracted from the vector of records.