records: add RDataBase::from trait method
The purpose is to give every record data conversion from RecordData
enum. This is an internal-only functionality added temporarily until
RecordsReader is implemented. Then implementation of Answer::from_msg
will be rewriten and record data will be returned directly from the
reader, instead of moving it out of ResourceRecord enum.