resolver_config: implement the Default trait
From now on ResolverConfig implements the Default trait.
By default the nameserver is not specified.
If a resolver receives a configuration with no specified nameserver,
creation of the resolver fails with NoNameservers error.

Note that ResolverConfig doesn't use an Option<SocketAddr> for
nameserver. This is possible because the UNSPECIFIED IpAddr can be used
as a default value, which denotes no nameserver scenario. This way the
API remains very convenient to use.

Implementation of the Default trait allows more ergonomic handling
of ResolverConfig in ch4.