message_reader: lazy calculation of Authority and Addition offsets
Calculation of the Answer offset is needed in vast majority of
cases. So, is very convenient to pre-calculate it in the constructor,
as it makes the `records()` and `records_reader()` methods infallible.

However, calculation of later offsets is not always needed, hence better
be calculated lazyly to not pay the price in the constructor.