cargo: use default requirement strategy for `tera` instead of tilde
There is some unexplained behavior (see [1]) in cargo when trying to bump
the tera dependency in ch4 to v1.16.0. I want to ensure that it is not
related to tilde dependency used in rsdns.

[1]

error: failed to select a version for `tera`.
    ... required by package `rsdns v0.13.0`
    ... which satisfies dependency `rsdns = "^0.13.0"` of package `ch4 v0.10.1 (/tmp/cargo-outdatedY7fi9e)`
versions that meet the requirements `~1.15.0` are: 1.15.0

all possible versions conflict with previously selected packages.

  previously selected package `tera v1.16.0`
    ... which satisfies dependency `tera = "^1.16.0"` of package `ch4 v0.10.1 (/tmp/cargo-outdatedY7fi9e)`

failed to select a version for `tera` which could resolve this conflict