cargo: enable default features in `tera`
There is some unexplained behavior in cargo updated when trying to bump
the tera requirement in ch4 to v1.16.0. This commit tries to use the
simplest form of requirement to check if this is the root cause.

[1]

error: failed to select a version for `tera`.
    ... required by package `rsdns v0.13.0`
    ... which satisfies dependency `rsdns = "^0.13.0"` of package `ch4 v0.10.1 (/tmp/cargo-outdatedY7fi9e)`
versions that meet the requirements `~1.15.0` are: 1.15.0

all possible versions conflict with previously selected packages.

  previously selected package `tera v1.16.0`
    ... which satisfies dependency `tera = "^1.16.0"` of package `ch4 v0.10.1 (/tmp/cargo-outdatedY7fi9e)`

failed to select a version for `tera` which could resolve this conflict