fuzz: temporarily remove generation of fuzz's coverage report
The stage `generate coverage.html` fails for quite some time now with
error [1]. It seems like the `llvm-cov` tool was changed. I haven't
found yet a way to make it work. So, this commit removes the code needed
for generation of the coverage report. The fuzz stage is still being
executed as before.

[1] error: Failed to load coverage: 'fuzz/target/x86_64-unknown-linux-gnu/release/rrset_from_msg': No coverage data found